Changes made while testing Tags, Users and Questions API

+++ src/main/java/com/forum/discussion_platform/dto/response/TagQuestionCountResponseDTO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class TagQuestionCountResponseDTO {
     private Long tagId;
     private String tagName;
-    private int questionCount;
+    private Long questionCount;
 }
